added optional password confirmation at mail creat
adds a field at account creation with email which asks to retype the
password to confirm it.
If the two password fields are differents, an error message pops up, and
the "_connexion" method is left.

this option is activated at SignInScreen creation,
with optional boolean parameter signUpPasswordCheck
(defaults to false, leaving the UI like it was before)

added 2 locale strings for "Confirm password" and "Passwords are
incorrect" messages.
I am not sure of the German translation, as I created it with Google Translate...